+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/tls/CustomX509TrustManagerBuilder.kt
@@ -13,7 +13,14 @@ class CustomX509TrustManagerBuilder(
     /**
      * CA 証明書を指定します。
      */
-    private val caCertificate: X509Certificate
+    private val caCertificate: X509Certificate,
+
+    /**
+     * すべての証明書を信頼するかどうかを指定します。
+     *
+     * デフォルトは false です。
+     */
+    private val insecure: Boolean = false,
 ) {
     /**
      * CA 証明書を使用して TLS 接続を行うためのカスタムされた TrustManager を構築します。
@@ -21,6 +28,15 @@ class CustomX509TrustManagerBuilder(
      * @return 指定された CA 証明書を使用する X509TrustManager
      */
     fun build(): X509TrustManager {
+        // insecure が true の場合、すべての証明書を信頼する TrustManager を返す
+        if (insecure) {
+             return object : X509TrustManager {
+                override fun checkClientTrusted(chain: Array<X509Certificate>, authType: String) = Unit
+                override fun checkServerTrusted(chain: Array<X509Certificate>, authType: String) = Unit
+                override fun getAcceptedIssuers(): Array<X509Certificate> = emptyArray()
+            }
+        }
+
         // 空の KeyStore を用意し、指定された CA 証明書を登録
         val keyStore = KeyStore.getInstance(KeyStore.getDefaultType()).apply {
             load(null, null)